requirements.txt
flask
python-dotenv
requests
pymongo
